+++ crypto/tss/ecdsa/cggmp/dkg/2_verify_handler.go
@@ -19,10 +19,10 @@ import (
 	"errors"
 	"math/big"
 
-	"github.com/getamis/alice/crypto/cggmp"
 	"github.com/getamis/alice/crypto/commitment"
 	"github.com/getamis/alice/crypto/ecpointgrouplaw"
 	"github.com/getamis/alice/crypto/tss"
+	"github.com/getamis/alice/crypto/tss/ecdsa/cggmp"
 	"github.com/getamis/alice/crypto/utils"
 	"github.com/getamis/alice/crypto/zkproof"
 	"github.com/getamis/alice/internal/message/types"
