+++ crypto/tss/ecdsa/cggmp/refresh/message.proto
@@ -16,7 +16,7 @@ syntax = "proto3";
 
 package refresh;
 
-option go_package = "github.com/getamis/alice/crypto/cggmp/refresh";
+option go_package = "github.com/getamis/alice/crypto/tss/ecdsa/cggmp/refresh";
 
 import "github.com/getamis/alice/crypto/ecpointgrouplaw/point.proto";
 import "github.com/getamis/alice/crypto/commitment/message.proto";
