+++ crypto/tss/ecdsa/cggmp/refresh/round_3.go
@@ -18,10 +18,10 @@ import (
 	"errors"
 	"math/big"
 
-	"github.com/getamis/alice/crypto/cggmp"
 	"github.com/getamis/alice/crypto/commitment"
 	pt "github.com/getamis/alice/crypto/ecpointgrouplaw"
 	"github.com/getamis/alice/crypto/tss"
+	"github.com/getamis/alice/crypto/tss/ecdsa/cggmp"
 	paillierzkproof "github.com/getamis/alice/crypto/zkproof/paillier"
 	"github.com/getamis/alice/internal/message/types"
 	"github.com/getamis/sirius/log"
