+++ crypto/tss/ecdsa/cggmp/sign/sign.go
