+++ crypto/tss/ecdsa/cggmp/signSix/message.proto
@@ -16,7 +16,7 @@ syntax = "proto3";
 
 package signSix;
 
-option go_package = "github.com/getamis/alice/crypto/cggmp/signSix";
+option go_package = "github.com/getamis/alice/crypto/tss/ecdsa/cggmp/signSix";
 
 import "github.com/getamis/alice/crypto/ecpointgrouplaw/point.proto";
 import "github.com/getamis/alice/crypto/zkproof/paillier/message.proto";
