+++ crypto/tss/ecdsa/cggmp/signSix/sign.go
